+++ ext/Protobuf_Ext.cmake
@@ -16,12 +16,15 @@ ExternalProject_Add(protobuf
 
 # get source dir after download step
 ExternalProject_Get_Property(protobuf SOURCE_DIR)
-add_custom_target(mprotobuf ALL
-      COMMAND git submodule update --init --recursive && ./autogen.sh && ./configure --prefix=${SOURCE_DIR} && make install -j8
-      WORKING_DIRECTORY ${SOURCE_DIR}
-      DEPENDS protobuf)
+if(NOT EXISTS ${SOURCE_DIR}/bin/protoc)
+    add_custom_target(mprotobuf ALL
+        COMMAND git submodule update --init --recursive && ./autogen.sh && ./configure --prefix=${SOURCE_DIR} && make install -j8
+        WORKING_DIRECTORY ${SOURCE_DIR}
+        DEPENDS protobuf)
+endif()
 
 set(PROTOBUF_INCLUDE_DIRS ${SOURCE_DIR}/include)
+file(MAKE_DIRECTORY ${PROTOBUF_INCLUDE_DIRS})
 set(PROTOBUF_LIBRARY_PATH ${SOURCE_DIR}/lib/libprotobuf.a)
 set(PROTOBUF_LIBRARY libpb)
 add_library(${PROTOBUF_LIBRARY} UNKNOWN IMPORTED)
